requirements.txt
# =========================
# Core Data Science
# =========================
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.11.0
scikit-learn>=1.3

# =========================
# Machine Learning Models
# =========================
xgboost>=2.0
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.16
einops>=0.7.0
accelerate>=0.27.0
transformers>=4.46.0
sentencepiece>=0.2.0

# =========================
# MLflow
# =========================
mlflow>=2.9

# =========================
# API
# =========================
fastapi>=0.104
uvicorn[standard]>=0.24
python-multipart>=0.0.7,<1.0
google-search-results>=2.4.2      
ddgs 


# =========================
# Serialization
# =========================
joblib>=1.3

# =========================
# Config & Utilities
# =========================
python-dotenv>=1.0.1
pydantic>=2.4
typing_extensions>=4.10.0
xmltodict>=0.13.0
babel>=2.15.0
requests>=2.31
loguru>=0.7

# =========================
# Computer Vision
# =========================
opencv-python-headless>=4.9.0
pillow>=10.3.0
easyocr
layoutparser
pytesseract

# =========================
# HuggingFace / LLM
# =========================
huggingface_hub>=0.23.0
hf_xet
langchain-google-genai==3.0.1
langchain >= 1.00

# =========================
# Visualization / Notebooks
# =========================
seaborn>=0.13.0
notebook>=7.0
jupyterlab>=4.1.0

# =========================
# Frontend
# =========================
streamlit>=1.33.0,<2.0
